Feat: 리뷰 작성 및 조회 기능 구현

+++ src/main/java/com/needle/FsoFso/product/service/ProductService.java
@@ -1,8 +1,8 @@
 package com.needle.FsoFso.product.service;
 
-import com.needle.FsoFso.product.dto.CartDto;
 import java.util.List;
 
+import com.needle.FsoFso.product.dto.CartDto;
 import com.needle.FsoFso.product.dto.ProductDto;
 
 public interface ProductService {
@@ -12,8 +12,8 @@ public interface ProductService {
 	List<ProductDto> getproducPagelist(int start);
 	
 	int getAllProductCount();
-
-	ProductDto getProductById(int id);
-
+	
+	ProductDto getProductById(long id);
+	
 	void addCart(CartDto cart);
 }
